fix(security): allow to run dev mode as any user
Add a HOME variable and writable directory such that dev mode can write
the root token to $HOME/.vault-token. Move set-capabilities call to
image creation time as an arbitrary user doesn't have the permission
to modify the capabilities of /bin/vault which belongs to root. This
means however that the image can not run without adding cap-add.